Refactor Sleep UI sections
