+++ android/app/src/main/java/com/noop/ui/SleepMetricCardsUi.kt
@@ -1,0 +1,592 @@
+package com.noop.ui
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.IntrinsicSize
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.heightIn
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.drawBehind
+import androidx.compose.ui.geometry.CornerRadius
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.StrokeCap
+import androidx.compose.ui.graphics.drawscope.DrawScope
+import androidx.compose.ui.semantics.contentDescription
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import com.noop.R
+import com.noop.analytics.SleepDebtLedger
+import java.util.Locale
+import kotlin.math.abs
+import kotlin.math.max
+
+// MARK: - 2. Metric grid (row-equalized min-height tiles, each with a bottom sparkline)
+
+@Composable
+internal fun MetricGrid(m: SleepModel, onMetricClick: (String) -> Unit = {}) {
+    val tiles = listOf<@Composable (Modifier) -> Unit>(
+        { mod ->
+            SparkTile(
+                mod, "Rest",
+                value = pctValue(m.performance.latest),
+                caption = vsTypical(m.performance.latest, m.performance.typical, "%"),
+                accent = m.performance.latest?.let { Palette.recoveryColor(it) } ?: Palette.textPrimary,
+                spark = m.performance.series, sparkColor = Palette.restColor,
+                onClick = { onMetricClick("performance") },
+            )
+        },
+        { mod ->
+            SparkTile(
+                mod, "Efficiency",
+                value = pctValue(m.efficiency.latest),
+                caption = vsTypical(m.efficiency.latest, m.efficiency.typical, "%"),
+                accent = Palette.statusPositive,
+                spark = m.efficiency.series, sparkColor = Palette.statusPositive,
+                onClick = { onMetricClick("efficiency") },
+            )
+        },
+        { mod ->
+            SparkTile(
+                mod, "Consistency",
+                value = pctValue(m.consistency.latest),
+                caption = vsTypical(m.consistency.latest, m.consistency.typical, "%"),
+                accent = m.consistency.latest?.let { Palette.recoveryColor(it) } ?: Palette.textPrimary,
+                spark = m.consistency.series, sparkColor = Palette.metricCyan,
+                onClick = { onMetricClick("consistency") },
+            )
+        },
+        { mod ->
+            SparkTile(
+                mod, "Hours vs Needed",
+                value = pctValue(m.hoursVsNeeded.latest),
+                caption = vsTypical(m.hoursVsNeeded.latest, m.hoursVsNeeded.typical, "%"),
+                accent = m.hoursVsNeeded.latest?.let { Palette.recoveryColor(minOf(100.0, it)) } ?: Palette.textPrimary,
+                spark = m.hoursVsNeeded.series, sparkColor = Palette.restColor,
+                onClick = { onMetricClick("hours_vs_needed") },
+            )
+        },
+        { mod ->
+            SparkTile(
+                mod, "Restorative",
+                value = pctValue(m.restorative.latest),
+                caption = vsTypical(m.restorative.latest, m.restorative.typical, "%"),
+                accent = Palette.sleepREM,
+                spark = m.restorative.series, sparkColor = Palette.sleepREM,
+                onClick = { onMetricClick("restorative") },
+            )
+        },
+        { mod ->
+            SparkTile(
+                mod, "Respiratory",
+                value = m.respiratory.latest?.let { String.format(Locale.US, "%.1f", it) } ?: "—",
+                caption = vsTypical(m.respiratory.latest, m.respiratory.typical, " rpm", decimals = 1),
+                accent = Palette.metricPurple,
+                spark = m.respiratory.series, sparkColor = Palette.metricPurple,
+                onClick = { onMetricClick("respiratory") },
+            )
+        },
+    )
+
+    Column(verticalArrangement = Arrangement.spacedBy(Metrics.gap)) {
+        SectionHeader("Night detail", overline = "Metrics", trailing = "vs typical")
+
+        // Sleep Debt is the actionable summary for the section, so it leads at the full
+        // two-column width. The remaining six peer metrics keep the established 2 × 3 grid.
+        SparkTile(
+            Modifier.fillMaxWidth(), "Sleep Debt",
+            value = m.sleepDebt.latest?.let { durationText(it) } ?: "—",
+            caption = debtCaption(m.sleepDebt.latest),
+            accent = debtColor(m.sleepDebt.latest),
+            spark = m.sleepDebt.series, sparkColor = Palette.metricRose,
+            onClick = { onMetricClick("sleep_debt") },
+        )
+
+        // Two-up rows; IntrinsicSize.Max + fillMaxHeight keep row neighbors equal height even when
+        // large font scales grow one tile past the tileHeight floor. No empty cells.
+        tiles.chunked(2).forEach { rowTiles ->
+            Row(
+                modifier = Modifier.height(IntrinsicSize.Max),
+                horizontalArrangement = Arrangement.spacedBy(Metrics.gap),
+            ) {
+                rowTiles.forEach { it(Modifier.weight(1f).fillMaxHeight()) }
+            }
+        }
+    }
+}
+
+// MARK: - 2b. Sleep-debt ledger (rolling 14-night running balance)
+
+/**
+ * A running balance of (slept − personal need) across the recent fortnight, surfaced as one
+ * card: the net debt/surplus headline, a plain-English read, and a diverging bar of each
+ * night's delta (surplus above the centre line, deficit below). Honest: a simple accumulator
+ * — a surplus night offsets a deficit one — capped at 14 nights, no-data nights skipped.
+ * Mirrors the macOS SleepView sleepDebtLedger card section-for-section. (#242)
+ */
+@Composable
+internal fun SleepDebtLedgerCard(ledger: SleepDebtLedger) {
+    Column(verticalArrangement = Arrangement.spacedBy(Metrics.gap)) {
+        SectionHeader("Sleep-debt ledger", overline = "Last 14 nights", trailing = "running balance")
+        NoopCard(padding = Metrics.cardPadding, tint = Palette.restColor) {
+            if (ledger.nightCount == 0) {
+                Text(
+                    uiString(R.string.l10n_sleep_screen_no_nights_with_sleep_data_yet_fa71b6b3),
+                    style = NoopType.subhead,
+                    color = Palette.textTertiary,
+                )
+            } else {
+                Column(verticalArrangement = Arrangement.spacedBy(Metrics.space14)) {
+                    // Headline: net balance + the short tag (sleep debt / surplus / balanced).
+                    Row(verticalAlignment = Alignment.CenterVertically) {
+                        Text(
+                            debtHeadline(ledger),
+                            style = NoopType.tileValueLarge,
+                            color = debtBalanceColor(ledger),
+                            maxLines = 1,
+                            overflow = TextOverflow.Ellipsis,
+                            modifier = Modifier.weight(1f),
+                        )
+                        Text(
+                            debtTag(ledger),
+                            style = NoopType.captionNumber,
+                            color = debtBalanceColor(ledger),
+                        )
+                    }
+                    // Plain-English read.
+                    Text(
+                        debtRead(ledger),
+                        style = NoopType.subhead,
+                        color = Palette.textSecondary,
+                    )
+                    // Per-night diverging delta bars (surplus up, deficit down).
+                    DebtDeltaBars(ledger)
+                    SleepMetricsHairline()
+                    SleepChartFooter(
+                        listOf(
+                            "Balance" to debtSigned(ledger.balanceMin),
+                            "Per-night need" to durationText(ledger.needMin),
+                            "Nights" to "${ledger.nightCount}",
+                        ),
+                    )
+                }
+            }
+        }
+    }
+}
+
+/**
+ * The diverging per-night delta strip: each night a bar from the centre line — up (accent)
+ * for a surplus, down (rose) for a deficit — scaled to the largest |delta|.
+ */
+@Composable
+private fun DebtDeltaBars(ledger: SleepDebtLedger) {
+    val deltas = ledger.nights.map { it.deltaMin }
+    val scale = max(deltas.maxOfOrNull { abs(it) } ?: 1.0, 1.0)
+    val accentColor = Palette.accent
+    val deficitColor = Palette.metricRose
+    val centreColor = Palette.hairline
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(56.dp)
+            .semantics {
+                contentDescription =
+                    uiString(R.string.l10n_sleep_screen_per_night_sleep_balance_ledger_nightcount_f339d0ab, ledger.nightCount, debtSigned(ledger.balanceMin))
+            }
+            .drawBehind {
+                val n = max(deltas.size, 1)
+                val slot = size.width / n
+                val barW = max(2f, slot * 0.6f)
+                val midY = size.height / 2f
+                // Centre (zero) line.
+                drawLine(
+                    color = centreColor,
+                    start = Offset(0f, midY),
+                    end = Offset(size.width, midY),
+                    strokeWidth = 1f,
+                )
+                deltas.forEachIndexed { i, d ->
+                    val frac = (abs(d) / scale).toFloat().coerceIn(0f, 1f)
+                    val h = max(2f, frac * (midY - 2f))
+                    val cx = slot * i + slot / 2f
+                    // Surplus grows upward from the centre, deficit downward.
+                    val top = if (d >= 0.0) midY - h else midY
+                    drawRoundRect(
+                        color = if (d >= 0.0) accentColor else deficitColor,
+                        topLeft = Offset(cx - barW / 2f, top),
+                        size = Size(barW, h),
+                        cornerRadius = CornerRadius(2f, 2f),
+                    )
+                }
+            },
+    )
+}
+
+// MARK: - 3. Stages vs typical
+
+@Composable
+internal fun StagesVsTypical(m: SleepModel) {
+    val s = m.stages
+    Column(verticalArrangement = Arrangement.spacedBy(Metrics.gap)) {
+        SectionHeader("Stages vs typical", overline = "Selected night", trailing = "marker = your mean")
+        NoopCard(tint = Palette.restColor) {
+            Column(verticalArrangement = Arrangement.spacedBy(Metrics.space14)) {
+                StageRow("Deep", last = s.deep, typical = m.typicalDeepMin, color = Palette.sleepDeep)
+                SleepMetricsHairline()
+                StageRow("REM", last = s.rem, typical = m.typicalRemMin, color = Palette.sleepREM)
+                SleepMetricsHairline()
+                StageRow("Light", last = s.light, typical = m.typicalLightMin, color = Palette.sleepLight)
+            }
+        }
+    }
+}
+
+@Composable
+internal fun SleepMetricsHairline() {
+    Box(modifier = Modifier.fillMaxWidth().height(Metrics.divider).background(Palette.hairline))
+}
+
+/** One stage bar: last-night minutes filled, with a vertical marker at the typical mean. */
+@Composable
+private fun StageRow(label: String, last: Double, typical: Double?, color: Color) {
+    val scaleMax = max(last, typical ?: 0.0) * 1.18
+    val scale = if (scaleMax > 0.0) scaleMax else 1.0
+    val deltaText: String = run {
+        if (typical == null || typical <= 0.0) {
+            ""
+        } else {
+            val diff = last - typical
+            val sign = if (diff >= 0) "+" else "−"
+            "$sign${durationText(abs(diff))} vs typ"
+        }
+    }
+    Column(verticalArrangement = Arrangement.spacedBy(Metrics.space6)) {
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            Overline(label, modifier = Modifier.weight(1f))
+            Text(durationText(last), style = NoopType.captionNumber, color = Palette.textPrimary)
+            if (deltaText.isNotEmpty()) {
+                Text(
+                    deltaText,
+                    style = NoopType.footnote,
+                    color = if (last >= (typical ?: last)) Palette.statusPositive else Palette.statusWarning,
+                    modifier = Modifier.padding(start = Metrics.space8),
+                )
+            }
+        }
+        // Track + last-night fill + typical marker.
+        val fillFrac = (last / scale).coerceIn(0.0, 1.0).toFloat()
+        val markerFrac = typical?.takeIf { it > 0.0 }?.let { (it / scale).coerceIn(0.0, 1.0).toFloat() }
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(Metrics.progressHeight)
+                .clip(RoundedCornerShape(Metrics.cornerPill))
+                .background(Palette.surfaceInset)
+                .semantics { contentDescription = uiString(R.string.l10n_sleep_screen_label_minutes_vs_your_typical_bar_b8f6a482, label) }
+                .drawBehind {
+                    // last-night fill
+                    if (fillFrac > 0f) {
+                        drawRoundRectFill(color, fillFrac)
+                    }
+                    // typical marker
+                    if (markerFrac != null) {
+                        val x = (size.width * markerFrac).coerceIn(1f, size.width - 1f)
+                        drawLine(
+                            color = Palette.textPrimary,
+                            start = Offset(x, 0f),
+                            end = Offset(x, size.height),
+                            strokeWidth = 2f,
+                            cap = StrokeCap.Round,
+                        )
+                    }
+                },
+        )
+    }
+}
+
+private fun DrawScope.drawRoundRectFill(color: Color, frac: Float) {
+    val w = (size.width * frac).coerceAtLeast(size.height)
+    val r = size.height / 2f
+    drawRoundRect(
+        color = color,
+        size = Size(w, size.height),
+        cornerRadius = CornerRadius(r, r),
+    )
+}
+
+// MARK: - 4. 14-day asleep-hours trend
+
+@Composable
+internal fun DurationTrend(m: SleepModel) {
+    val pts = m.trendHours
+    val avg = pts.sleepAverageOrNull()
+    Column(verticalArrangement = Arrangement.spacedBy(Metrics.gap)) {
+        SectionHeader("Trend", overline = "Sleep", trailing = "Last 14 days")
+        SleepChartCard(
+            title = uiString(R.string.l10n_sleep_screen_hours_asleep_06f68993),
+            subtitle = "Per night, trailing 14 days",
+            trailing = avg?.let { String.format(Locale.US, "%.1f h avg", it) },
+            tint = Palette.restColor,
+            footer = {
+                SleepChartFooter(
+                    listOf(
+                        "Avg" to (avg?.let { String.format(Locale.US, "%.1f h", it) } ?: "—"),
+                        "Min" to (pts.minOrNull()?.let { String.format(Locale.US, "%.1f h", it) } ?: "—"),
+                        "Max" to (pts.maxOrNull()?.let { String.format(Locale.US, "%.1f h", it) } ?: "—"),
+                        "Nights" to "${pts.size}",
+                    ),
+                )
+            },
+        ) {
+            if (pts.size >= 2) {
+                Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
+                    // #85: sleep duration reads as a per-night histogram (zero-based bars), matching the
+                    // iOS Sleep tab's TrendChart(showsBars:) — a BarMark is proportional to hours slept,
+                    // clearer than a line for a nightly total. BarChart floors at 0 like the iOS bar domain.
+                    BarChart(
+                        values = pts,
+                        modifier = Modifier.fillMaxWidth().height(Metrics.compactChartHeight)
+                            .semantics { contentDescription = uiString(R.string.l10n_sleep_screen_sleep_hours_trend_chart_a6fbc46d) },
+                        color = Palette.restColor,
+                        selectionEnabled = true,
+                        // #691: on tap, show the DATE alongside the value (the shared chart's tooltip),
+                        // matching the other trend graphs. trendDates is index-aligned with the values.
+                        selectionLabels = m.trendDates.map(::shortDayLabel),
+                    )
+                    DateAxisRow(m.trendDates)
+                }
+            } else {
+                TrendPlaceholder()
+            }
+        }
+
+        SleepChartCard(
+            title = uiString(R.string.l10n_sleep_screen_sleep_debt_3aec7d9c),
+            subtitle = "Sleep debt per day",
+            // #691: sleep debt is usually well under an hour, so decimal hours ("0.6h") reads badly —
+            // show hours+minutes. trendDebtHours is in hours; durationText takes minutes.
+            trailing = m.trendDebtHours.lastOrNull()?.let { durationText(it * 60.0) },
+            tint = Palette.restColor,
+            footer = {
+                SleepChartFooter(
+                    listOf(
+                        "Avg" to (m.trendDebtHours.sleepAverageOrNull()?.let { durationText(it * 60.0) } ?: "â€”"),
+                        "Max" to (m.trendDebtHours.maxOrNull()?.let { durationText(it * 60.0) } ?: "â€”"),
+                        "Days" to "${m.trendDebtHours.size}",
+                    ),
+                )
+            },
+        ) {
+            if (m.trendDebtHours.isNotEmpty()) {
+                Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
+                    BarChart(
+                        values = m.trendDebtHours,
+                        modifier = Modifier.fillMaxWidth().height(Metrics.compactChartHeight)
+                            .semantics { contentDescription = uiString(R.string.l10n_sleep_screen_sleep_debt_trend_chart_9e178776) },
+                        color = Palette.metricRose,
+                        selectionEnabled = true,
+                        selectionLabels = m.trendDates.map(::shortDayLabel),   // #691: hover shows date + value
+                    )
+                    DateAxisRow(m.trendDates)
+                }
+            } else {
+                TrendPlaceholder()
+            }
+        }
+    }
+}
+
+@Composable
+private fun TrendPlaceholder() {
+    Box(
+        modifier = Modifier.fillMaxWidth(),
+        contentAlignment = Alignment.Center,
+    ) {
+        InsetChartPlaceholder(message = "Not enough nights yet.")
+    }
+}
+
+@Composable
+private fun TrendLegend(items: List<Pair<String, Color>>) {
+    Row(horizontalArrangement = Arrangement.spacedBy(Metrics.space14)) {
+        items.forEach { (label, color) ->
+            Row(
+                horizontalArrangement = Arrangement.spacedBy(Metrics.space6),
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Box(
+                    modifier = Modifier
+                        .width(Metrics.legendLineWidth)
+                        .height(Metrics.legendLineHeight)
+                        .clip(RoundedCornerShape(Metrics.cornerPill))
+                        .background(color),
+                )
+                Text(label, style = NoopType.footnote, color = Palette.textTertiary)
+            }
+        }
+    }
+}
+
+@Composable
+private fun DateAxisRow(days: List<String>) {
+    if (days.isEmpty()) return
+    val labels = listOf(
+        days.firstOrNull(),
+        days.getOrNull(days.lastIndex / 2),
+        days.lastOrNull(),
+    ).map { it?.let(::shortDayLabel).orEmpty() }
+    Row(modifier = Modifier.fillMaxWidth()) {
+        labels.forEach { label ->
+            Text(
+                text = label,
+                style = NoopType.footnote,
+                color = Palette.textTertiary,
+                modifier = Modifier.weight(1f),
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+            )
+        }
+    }
+}
+
+// MARK: - ChartCard / ChartFooter (local — mirror the macOS ChartCard the screen used)
+
+/**
+ * The chart container the macOS screen leaned on: a NoopCard with a header (overline-
+ * style title + subtitle + trailing read-out), the chart body, then a footer row of
+ * label/value pairs. Kept local so the shared component set stays minimal.
+ */
+@Composable
+internal fun SleepChartCard(
+    title: String,
+    subtitle: String,
+    trailing: String?,
+    footer: @Composable () -> Unit,
+    tint: Color? = null,
+    chart: @Composable () -> Unit,
+) {
+    NoopCard(padding = Metrics.cardPadding, tint = tint) {
+        Column(verticalArrangement = Arrangement.spacedBy(Metrics.space14)) {
+            Row(verticalAlignment = Alignment.Top) {
+                Column(modifier = Modifier.weight(1f)) {
+                    Text(title, style = NoopType.headline, color = Palette.textPrimary)
+                    Text(subtitle, style = NoopType.footnote, color = Palette.textSecondary)
+                }
+                if (trailing != null) {
+                    Text(trailing, style = NoopType.chartValue, color = Palette.textPrimary)
+                }
+            }
+            chart()
+            footer()
+        }
+    }
+}
+
+/** A footer strip of label/value pairs, evenly distributed. */
+@Composable
+private fun SleepChartFooter(items: List<Pair<String, String>>) {
+    Row(modifier = Modifier.fillMaxWidth()) {
+        items.forEach { (label, value) ->
+            Column(modifier = Modifier.weight(1f)) {
+                Overline(label, color = Palette.textTertiary)
+                // Stage-breakdown values like "1h 23m (24%)" wrapped to a second line in a narrow column,
+                // pushing the row taller and clipping against the card edge (#406). Hold them to one line.
+                Text(
+                    value,
+                    style = NoopType.captionNumber,
+                    color = Palette.textPrimary,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis,
+                    softWrap = false,
+                )
+            }
+        }
+    }
+}
+
+// MARK: - SparkTile (min-height metric tile, stacked: value + caption over a full-width 30-day sparkline)
+
+@Composable
+private fun SparkTile(
+    modifier: Modifier,
+    label: String,
+    value: String,
+    caption: String?,
+    accent: Color,
+    spark: List<Double>,
+    sparkColor: Color,
+    onClick: (() -> Unit)? = null,
+) {
+    // liquidPress on the tappable tile: it settles inward on press (the pilot's card feel). The SAME
+    // interactionSource drives the clickable + the press; indication = null so only the liquid settle shows.
+    val interaction = remember { MutableInteractionSource() }
+    // heightIn (not height): tileHeight is a floor, matching the Swift StatTile. At normal font scale the
+    // tile keeps its 108dp footprint; at large font scales it grows instead of clipping the caption. (#squish)
+    val clickMod = if (onClick != null) {
+        modifier
+            .heightIn(min = Metrics.tileHeight)
+            .liquidPress(interaction)
+            .clickable(interactionSource = interaction, indication = null, onClick = onClick)
+    } else {
+        modifier.heightIn(min = Metrics.tileHeight)
+    }
+    NoopCard(modifier = clickMod, padding = Metrics.space14) {
+        // fillMaxHeight so the weight-spacer can pin the sparkline to the card bottom once the
+        // MetricGrid row bounds the height (Row height(IntrinsicSize.Max) + tile fillMaxHeight()).
+        Column(modifier = Modifier.fillMaxWidth().fillMaxHeight()) {
+            Overline(label)
+            Text(
+                value,
+                style = NoopType.tileValue,
+                color = accent,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+            )
+            if (caption != null) {
+                Text(
+                    caption,
+                    style = NoopType.footnote,
+                    color = Palette.textTertiary,
+                    // Full card width now, so the "-3% vs typical" caption fits; ellipsis stays as a
+                    // safety net for extreme localized strings.
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis,
+                    modifier = Modifier.padding(top = Metrics.space2),
+                )
+            }
+            Spacer(Modifier.weight(1f))
+            val tail = spark.takeLast(30)
+            if (tail.size >= 2) {
+                // Full-width bottom spark. Outer height(sparkHeight) deliberately overrides Sparkline's
+                // internal 28dp default down to the 22dp tile spark (same override SparkTailBox does).
+                Sparkline(
+                    values = tail,
+                    color = sparkColor,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(top = Metrics.space8)
+                        .height(Metrics.sparkHeight),
+                )
+            }
+        }
+    }
+}
