+++ android/app/src/main/java/com/noop/ui/SleepScreen.kt
@@ -1010,7 +1010,7 @@ private fun Hero(
             // else keeps the honest proportional strip + StageBreakdownRows footer.
             val real = display.realSegments?.takeIf { it.size >= 2 }
             if (real != null) {
-                ChartCard(
+                SleepChartCard(
                     title = uiString(R.string.l10n_sleep_screen_stage_breakdown_e9b714f9),
                     subtitle = subtitle,
                     trailing = durationText(s.asleep),
@@ -1029,7 +1029,7 @@ private fun Hero(
                     )
                 }
             } else {
-                ChartCard(
+                SleepChartCard(
                     title = uiString(R.string.l10n_sleep_screen_stage_breakdown_e9b714f9),
                     subtitle = subtitle,
                     trailing = durationText(s.asleep),
@@ -1640,1057 +1640,9 @@ private fun MotionStrip(epochs: List<Double>) {
     }
 }
 
-/**
- * "Asleep / Woke" — the fell-asleep and woke clock times for the navigated night, read off the
- * session's onset (startTs) and wake (endTs) timestamps, each with a moon / sun glyph. Sits in the
- * hero between the night-nav header and the stage card so the two times people glance for first are
- * always visible, not truncated in the header caption. On-brand (surfaceRaised block, tokens) and
- * combined into one TalkBack element. Mirrors iOS SleepView.sleepWindowRow (PR #289).
- */
-@Composable
-private fun SleepWindowRow(onsetTs: Long, wakeTs: Long) {
-    val asleep = clockTimeLabel(onsetTs)
-    val woke = clockTimeLabel(wakeTs)
-    // A frosted Rest-tinted card (was a flat surfaceRaised block) so the window row sits in the
-    // same colour world as the rest of the screen. Bevel treatment — content unchanged.
-    NoopCard(
-        modifier = Modifier.semantics(mergeDescendants = true) {
-            contentDescription = uiString(R.string.l10n_sleep_screen_fell_asleep_at_asleep_woke_at_80465b2d, asleep, woke)
-        },
-        padding = Metrics.space14,
-        tint = Palette.restColor,
-    ) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            SleepTime(icon = Icons.Filled.Bedtime, label = uiString(R.string.l10n_sleep_screen_asleep_b9692bbe), value = asleep)
-            Spacer(Modifier.width(Metrics.space12))
-            Box(
-                modifier = Modifier
-                    .height(30.dp)
-                    .width(Metrics.divider)
-                    .background(Palette.hairline),
-            )
-            Spacer(Modifier.width(Metrics.space12))
-            SleepTime(icon = Icons.Filled.WbSunny, label = uiString(R.string.l10n_sleep_screen_woke_cfbb59a8), value = woke)
-            Spacer(Modifier.weight(1f))
-        }
-    }
-}
+// MARK: - Sleep window and night navigation UI lives in SleepNightNavUi.kt
 
-@Composable
-private fun SleepTime(icon: androidx.compose.ui.graphics.vector.ImageVector, label: String, value: String) {
-    Row(
-        horizontalArrangement = Arrangement.spacedBy(Metrics.space10),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Icon(
-            icon,
-            contentDescription = null, // row carries the combined description
-            tint = Palette.restColor,
-            modifier = Modifier.size(20.dp),
-        )
-        Column(verticalArrangement = Arrangement.spacedBy(Metrics.space2)) {
-            Overline(label, color = Palette.textTertiary)
-            Text(value, style = NoopType.number(22f), color = Palette.textPrimary, maxLines = 1)
-        }
-    }
-}
-
-/**
- * Hero header with ◀/▶ to browse past nights plus an accent-tinted center block that
- * mirrors the Today page's date-nav: tapping the block opens a [DatePickerDialog] to jump
- * to any night by date, and the edit-pen icon opens a chooser to adjust the session's
- * bed/wake times via [TimePickerDialog]. ◀ goes older (offset+1), ▶ newer; each is disabled
- * at its bound — tinted tertiary when disabled, accent when active. (#160)
- */
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun NightNavHeader(
-    offset: Int,
-    lastIndex: Int,
-    clock: String?,
-    onNavigate: (Int) -> Unit,
-    session: SleepSession? = null,
-    onUpdateTimes: (SleepSession, Long, Long) -> Unit = { _, _, _ -> },
-    onDeleteSession: (SleepSession) -> Unit = {},
-    onAddNap: (Long, Long) -> Unit = { _, _ -> },
-    onPickNightDate: ((LocalDate) -> Unit)? = null,
-) {
-    val canGoOlder = offset < lastIndex
-    val canGoNewer = offset > 0
-    val context = LocalContext.current
-    var showTimeChoice by remember { mutableStateOf(false) }
-    var showDeleteConfirm by remember { mutableStateOf(false) }
-    var editingBed by remember { mutableStateOf(false) }
-    var editingWake by remember { mutableStateOf(false) }
-    var sleepEditDraft by remember(session?.deviceId, session?.startTs) {
-        mutableStateOf<SleepTimeEditDraft?>(null)
-    }
-    var showDatePicker by remember { mutableStateOf(false) }
-    // #940 guard 2: a corrected (start, end) window that no longer touches the night's recorded
-    // coverage parks here awaiting an explicit confirm; committing it silently fabricated an
-    // all-awake phantom night that hid the tab's history. null = nothing pending.
-    var pendingDisjointTimes by remember { mutableStateOf<Pair<Long, Long>?>(null) }
-    // Manual nap add (#508): pick a start time, then an end time; both anchored to THIS night's wake day
-    // so the new nap lands on the right day. napStartTs holds the chosen start between the two pickers.
-    var addingNapStart by remember { mutableStateOf(false) }
-    var addingNapEnd by remember { mutableStateOf(false) }
-    var napStartTs by remember { mutableStateOf(0L) }
-
-    // Commit funnel for the COMPLETE drafted window (#515/#940). Neither picker writes by itself:
-    // only Save reaches this function, so an edited bedtime can never be persisted against the old
-    // wake (or vice versa). A window outside the recorded coverage still uses #940's explicit confirm.
-    fun commitTimes(s: SleepSession, newStart: Long, newEnd: Long) {
-        val coverageStart = minOf(s.startTs, s.effectiveStartTs)
-        if (SleepEditGuard.isDisjoint(newStart, newEnd, coverageStart, s.endTs)) {
-            pendingDisjointTimes = newStart to newEnd
-        } else {
-            onUpdateTimes(s, newStart, newEnd)
-        }
-    }
-
-    // Atomic editor (#515): both rows mutate an in-memory draft. Save validates and commits the pair
-    // once; Cancel discards it. This mirrors Apple SleepTimeEditor's single start+end save funnel.
-    val currentDraft = sleepEditDraft
-    if (showTimeChoice && session != null && currentDraft != null) {
-        val timeFmt = SimpleDateFormat("HH:mm", Locale.US)
-        val bedText = timeFmt.format(Date(currentDraft.startTs * 1000L))
-        val wakeText = timeFmt.format(Date(currentDraft.endTs * 1000L))
-        val validated = currentDraft.validatedWindow(System.currentTimeMillis() / 1000L)
-        val blockShape2 = RoundedCornerShape(Metrics.cornerSm)
-        androidx.compose.material3.AlertDialog(
-            onDismissRequest = {
-                showTimeChoice = false
-                sleepEditDraft = null
-            },
-            containerColor = Palette.surfaceRaised,
-            titleContentColor = Palette.textPrimary,
-            textContentColor = Palette.textSecondary,
-            title = { Text(uiString(R.string.l10n_sleep_screen_adjust_sleep_times_1e325561), style = NoopType.headline) },
-            text = {
-                Column(verticalArrangement = Arrangement.spacedBy(Metrics.space6)) {
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clip(blockShape2)
-                            .background(Palette.surfaceOverlay)
-                            .clickable { showTimeChoice = false; editingBed = true }
-                            .padding(horizontal = Metrics.space16, vertical = Metrics.space14),
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        Column(modifier = Modifier.weight(1f)) {
-                            Overline("Bedtime", color = Palette.textTertiary)
-                            Spacer(Modifier.height(Metrics.space4))
-                            Text(bedText, style = NoopType.headline, color = Palette.textPrimary)
-                        }
-                        Icon(Icons.Filled.Edit, contentDescription = null, tint = Palette.accent, modifier = Modifier.size(20.dp))
-                    }
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clip(blockShape2)
-                            .background(Palette.surfaceOverlay)
-                            .clickable { showTimeChoice = false; editingWake = true }
-                            .padding(horizontal = Metrics.space16, vertical = Metrics.space14),
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        Column(modifier = Modifier.weight(1f)) {
-                            Overline("Wake-up", color = Palette.textTertiary)
-                            Spacer(Modifier.height(Metrics.space4))
-                            Text(wakeText, style = NoopType.headline, color = Palette.textPrimary)
-                        }
-                        Icon(Icons.Filled.Edit, contentDescription = null, tint = Palette.accent, modifier = Modifier.size(20.dp))
-                    }
-                }
-            },
-            confirmButton = {
-                TextButton(
-                    enabled = validated != null,
-                    onClick = {
-                        val window = validated ?: return@TextButton
-                        showTimeChoice = false
-                        sleepEditDraft = null
-                        commitTimes(session, window.first, window.second)
-                    },
-                ) {
-                    Text(
-                        uiString(R.string.l10n_sleep_screen_save_efc007a3),
-                        style = NoopType.body,
-                        color = if (validated != null) Palette.accent else Palette.textTertiary,
-                    )
-                }
-            },
-            dismissButton = {
-                TextButton(onClick = {
-                    showTimeChoice = false
-                    sleepEditDraft = null
-                }) {
-                    Text(
-                        uiString(R.string.l10n_sleep_screen_cancel_77dfd213),
-                        style = NoopType.body,
-                        color = Palette.textSecondary,
-                    )
-                }
-            },
-        )
-    }
-
-    // Bed-time picker mutates only the draft. Returning to the parent dialog lets the user inspect and
-    // adjust BOTH endpoints before the single Save (#515). The cross-midnight correction stays in the
-    // pure SleepTimeEditDraft/SleepEditGuard path pinned by JVM tests.
-    val draftForBed = sleepEditDraft
-    if (editingBed && session != null && draftForBed != null) {
-        val startCal = Calendar.getInstance().apply { timeInMillis = draftForBed.startTs * 1000L }
-        DisposableEffect(Unit) {
-            val dialog = TimePickerDialog(
-                context,
-                { _, h, m ->
-                    val cal = Calendar.getInstance().apply {
-                        timeInMillis = draftForBed.startTs * 1000L
-                        set(Calendar.HOUR_OF_DAY, h); set(Calendar.MINUTE, m)
-                        set(Calendar.SECOND, 0); set(Calendar.MILLISECOND, 0)
-                    }
-                    sleepEditDraft = draftForBed.withBedCandidate(
-                        candidateBedTs = cal.timeInMillis / 1000L,
-                        nowTs = System.currentTimeMillis() / 1000L,
-                    )
-                },
-                startCal.get(Calendar.HOUR_OF_DAY),
-                startCal.get(Calendar.MINUTE),
-                true,
-            ).apply { setTitle("Bedtime") }
-            dialog.setOnDismissListener {
-                editingBed = false
-                if (sleepEditDraft != null) showTimeChoice = true
-            }
-            dialog.show()
-            onDispose { runCatching { dialog.dismiss() } }
-        }
-    }
-
-    // Wake-up picker also mutates only the draft. Its calendar day is derived from the DRAFT bedtime,
-    // so editing bedtime first and wake second produces one coherent cross-midnight window (#515/#406).
-    val draftForWake = sleepEditDraft
-    if (editingWake && session != null && draftForWake != null) {
-        val endCal = Calendar.getInstance().apply { timeInMillis = draftForWake.endTs * 1000L }
-        DisposableEffect(Unit) {
-            val dialog = TimePickerDialog(
-                context,
-                { _, h, m ->
-                    sleepEditDraft = draftForWake.withWakeTime(hour = h, minute = m)
-                },
-                endCal.get(Calendar.HOUR_OF_DAY),
-                endCal.get(Calendar.MINUTE),
-                true,
-            ).apply { setTitle("Wake-up time") }
-            dialog.setOnDismissListener {
-                editingWake = false
-                if (sleepEditDraft != null) showTimeChoice = true
-            }
-            dialog.show()
-            onDispose { runCatching { dialog.dismiss() } }
-        }
-    }
-
-    // Date jump — capped at today so a future night can't be selected.
-    if (showDatePicker && onPickNightDate != null) {
-        val cal = session?.let { Calendar.getInstance().apply { timeInMillis = it.effectiveStartTs * 1000L } }
-            ?: Calendar.getInstance()
-        DisposableEffect(Unit) {
-            val dialog = DatePickerDialog(
-                context,
-                { _, year, month, day ->
-                    onPickNightDate(LocalDate.of(year, month + 1, day))
-                    showDatePicker = false
-                },
-                cal.get(Calendar.YEAR),
-                cal.get(Calendar.MONTH),
-                cal.get(Calendar.DAY_OF_MONTH),
-            ).apply {
-                datePicker.maxDate = System.currentTimeMillis()
-                setOnDismissListener { showDatePicker = false }
-            }
-            dialog.show()
-            onDispose { runCatching { dialog.dismiss() } }
-        }
-    }
-
-    // Manual nap (#508) step 1: pick the nap's START time, anchored to the night's wake DAY (a natural
-    // place to look for a missed daytime nap). Defaults to ~1h after the night's wake.
-    if (addingNapStart && session != null) {
-        val anchorTs = session.endTs + 3_600L
-        val startCal = Calendar.getInstance().apply { timeInMillis = anchorTs * 1000L }
-        DisposableEffect(Unit) {
-            val dialog = TimePickerDialog(
-                context,
-                { _, h, m ->
-                    val cal = Calendar.getInstance().apply {
-                        timeInMillis = anchorTs * 1000L
-                        set(Calendar.HOUR_OF_DAY, h); set(Calendar.MINUTE, m)
-                        set(Calendar.SECOND, 0); set(Calendar.MILLISECOND, 0)
-                    }
-                    // #940: a nap being logged already happened. The anchor day is the night's wake
-                    // day (usually today), so a picked time later than the clock means the most
-                    // recent PAST occurrence: snap back a day (no wake rule here; a nap after the
-                    // night's wake is normal).
-                    napStartTs = SleepEditGuard.autoCorrectedBed(
-                        previousBedTs = anchorTs,
-                        candidateBedTs = cal.timeInMillis / 1000L,
-                        originalWakeTs = null,
-                        nowTs = System.currentTimeMillis() / 1000L,
-                    )
-                    addingNapStart = false
-                    addingNapEnd = true
-                },
-                startCal.get(Calendar.HOUR_OF_DAY),
-                startCal.get(Calendar.MINUTE),
-                true,
-            ).apply { setTitle("Nap started") }
-            dialog.setOnDismissListener { addingNapStart = false }
-            dialog.show()
-            onDispose { runCatching { dialog.dismiss() } }
-        }
-    }
-
-    // Manual nap (#508) step 2: pick the nap's END time — TIME-ONLY, its day DERIVED from the chosen start
-    // (first instant strictly after start, within 24h), mirroring the wake-edit cross-day constraint so a
-    // nap can't be re-bucketed onto the wrong day. Then hand (start, end) to onAddNap.
-    if (addingNapEnd && napStartTs > 0L) {
-        val endCal = Calendar.getInstance().apply { timeInMillis = (napStartTs + 30 * 60L) * 1000L }
-        DisposableEffect(Unit) {
-            val dialog = TimePickerDialog(
-                context,
-                { _, h, m ->
-                    val startTs = napStartTs
-                    val cal = Calendar.getInstance().apply {
-                        timeInMillis = startTs * 1000L
-                        set(Calendar.HOUR_OF_DAY, h); set(Calendar.MINUTE, m)
-                        set(Calendar.SECOND, 0); set(Calendar.MILLISECOND, 0)
-                        if (timeInMillis / 1000L <= startTs) add(Calendar.DAY_OF_MONTH, 1)
-                    }
-                    onAddNap(startTs, cal.timeInMillis / 1000L)
-                    addingNapEnd = false
-                    napStartTs = 0L
-                },
-                endCal.get(Calendar.HOUR_OF_DAY),
-                endCal.get(Calendar.MINUTE),
-                true,
-            ).apply { setTitle("Nap ended") }
-            dialog.setOnDismissListener { addingNapEnd = false }
-            dialog.show()
-            onDispose { runCatching { dialog.dismiss() } }
-        }
-    }
-
-    // #940 guard 2's consent step: the corrected window no longer touches the night's recorded
-    // coverage, so there is nothing to stage it from. Same wording as the iOS SleepTimeEditor alert.
-    val pendingTimes = pendingDisjointTimes
-    if (pendingTimes != null && session != null) {
-        androidx.compose.material3.AlertDialog(
-            onDismissRequest = { pendingDisjointTimes = null },
-            containerColor = Palette.surfaceRaised,
-            titleContentColor = Palette.textPrimary,
-            textContentColor = Palette.textSecondary,
-            title = { Text(uiString(R.string.l10n_sleep_screen_move_this_sleep_438dd3b5), style = NoopType.headline) },
-            text = {
-                Text(
-                    uiString(R.string.l10n_sleep_screen_this_moves_the_night_to_a_fac2fb46),
-                    style = NoopType.subhead,
-                )
-            },
-            confirmButton = {
-                TextButton(onClick = {
-                    onUpdateTimes(session, pendingTimes.first, pendingTimes.second)
-                    pendingDisjointTimes = null
-                }) { Text(uiString(R.string.l10n_sleep_screen_move_anyway_19ee824d), style = NoopType.subhead, color = Palette.statusWarning) }
-            },
-            dismissButton = {
-                TextButton(onClick = { pendingDisjointTimes = null }) {
-                    Text(uiString(R.string.l10n_sleep_screen_cancel_77dfd213), style = NoopType.subhead, color = Palette.textSecondary)
-                }
-            },
-        )
-    }
-
-    val nightLabel = nightRelativeLabel(offset)
-    val blockShape = RoundedCornerShape(Metrics.cornerSm)
-    val clockParts = clock?.split(" · ", limit = 2)
-    val dateLabel = clockParts?.getOrNull(0)
-    val timeLabel = clockParts?.getOrNull(1)
-
-    Column(verticalArrangement = Arrangement.spacedBy(Metrics.space6)) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy(Metrics.selectorSpacing),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            IconButton(onClick = { if (canGoOlder) onNavigate(offset + 1) }, enabled = canGoOlder) {
-                Icon(Icons.Filled.ChevronLeft, contentDescription = uiString(R.string.l10n_sleep_screen_previous_night_9f339047), tint = if (canGoOlder) Palette.accent else Palette.textTertiary)
-            }
-            Column(
-                modifier = Modifier
-                    .weight(1f)
-                    .clip(blockShape)
-                    // Clean material surface (matches DayNavBar) — no gold wash behind the date;
-                    // the gold pop lives only on the date text below.
-                    .background(Palette.surfaceInset)
-                    .border(Metrics.divider, Palette.hairline, blockShape)
-                    .clickable(enabled = onPickNightDate != null, onClickLabel = "Pick night date") { showDatePicker = true }
-                    .padding(vertical = Metrics.selectorPadding, horizontal = Metrics.selectorPadding),
-                horizontalAlignment = Alignment.CenterHorizontally,
-            ) {
-                Text(nightLabel, style = NoopType.caption, color = Palette.textPrimary, maxLines = 1, overflow = TextOverflow.Ellipsis)
-                if (dateLabel != null) {
-                    Text(dateLabel, style = NoopType.captionNumber, color = Palette.accentHover, maxLines = 1, overflow = TextOverflow.Ellipsis)
-                }
-            }
-            IconButton(onClick = { if (canGoNewer) onNavigate(offset - 1) }, enabled = canGoNewer) {
-                Icon(Icons.Filled.ChevronRight, contentDescription = uiString(R.string.l10n_sleep_screen_next_night_7deeb06b), tint = if (canGoNewer) Palette.accent else Palette.textTertiary)
-            }
-        }
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.Center,
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text(
-                timeLabel ?: clock ?: "—",
-                style = NoopType.captionNumber,
-                color = Palette.accent,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-            if (session != null) {
-                Spacer(Modifier.width(Metrics.space6))
-                Icon(
-                    Icons.Filled.Edit,
-                    contentDescription = uiString(R.string.l10n_sleep_screen_adjust_sleep_times_1e325561),
-                    tint = Palette.textTertiary,
-                    modifier = Modifier.size(14.dp).clickable {
-                        sleepEditDraft = SleepTimeEditDraft(session.effectiveStartTs, session.endTs)
-                        showTimeChoice = true
-                    },
-                )
-                Spacer(Modifier.width(Metrics.space12))
-                Icon(
-                    Icons.Filled.DeleteOutline,
-                    contentDescription = uiString(R.string.l10n_sleep_screen_delete_this_sleep_session_6932e931),
-                    tint = Palette.textTertiary,
-                    modifier = Modifier.size(14.dp).clickable { showDeleteConfirm = true },
-                )
-                // Add a missed nap as its OWN session (#508) — staged from raw, never folded into this
-                // night's main sleep. Two pickers (start → end), the end day derived from the start.
-                Spacer(Modifier.width(Metrics.space12))
-                Icon(
-                    Icons.Filled.Add,
-                    contentDescription = uiString(R.string.l10n_sleep_screen_add_a_nap_a1b3204f),
-                    tint = Palette.textTertiary,
-                    modifier = Modifier.size(14.dp).clickable { addingNapStart = true },
-                )
-            }
-        }
-        // When the older-night arrow is disabled because no earlier night is banked yet, the chevron
-        // just greying out reads as broken. Show a short, honest hint instead — earlier nights only
-        // appear once the strap has offloaded them (typically the next morning sync). (#614 follow-up)
-        if (!canGoOlder) {
-            Text(
-                uiString(R.string.l10n_sleep_screen_no_earlier_night_stored_yet_earlier_ab637d4c),
-                style = NoopType.footnote,
-                color = Palette.textTertiary,
-                textAlign = TextAlign.Center,
-                modifier = Modifier.fillMaxWidth(),
-            )
-        }
-    }
-
-    // Confirm before removing the night — the same on-brand AlertDialog the time-edit chooser
-    // uses (surfaceRaised, Noop type tokens), not a bare Material default. (#281)
-    if (showDeleteConfirm && session != null) {
-        androidx.compose.material3.AlertDialog(
-            onDismissRequest = { showDeleteConfirm = false },
-            containerColor = Palette.surfaceRaised,
-            titleContentColor = Palette.textPrimary,
-            textContentColor = Palette.textSecondary,
-            title = { Text(uiString(R.string.l10n_sleep_screen_delete_this_sleep_session_c347b909), style = NoopType.headline) },
-            text = {
-                // A detected night is tombstoned so it won't re-detect; a userEdited/nap row writes no
-                // tombstone, so its copy drops that (false) promise. Mirrors the undo banner. (#65)
-                Text(
-                    if (session.userEdited) {
-                        "Removes this sleep and recomputes the day without it. You can undo for a few seconds after."
-                    } else {
-                        "Removes this recorded sleep and recomputes the day without it. NOOP won't re-detect sleep in this window. You can undo for a few seconds after."
-                    },
-                    style = NoopType.subhead,
-                )
-            },
-            confirmButton = {
-                TextButton(onClick = {
-                    showDeleteConfirm = false
-                    onDeleteSession(session)
-                }) {
-                    Text(uiString(R.string.l10n_sleep_screen_delete_f6fdbe48), style = NoopType.headline, color = Palette.statusCritical)
-                }
-            },
-            dismissButton = {
-                TextButton(onClick = { showDeleteConfirm = false }) {
-                    Text(uiString(R.string.l10n_sleep_screen_cancel_77dfd213), style = NoopType.subhead, color = Palette.textTertiary)
-                }
-            },
-        )
-    }
-}
-
-// MARK: - 2. Metric grid (row-equalized min-height tiles, each with a bottom sparkline)
-
-@Composable
-private fun MetricGrid(m: SleepModel, onMetricClick: (String) -> Unit = {}) {
-    val tiles = listOf<@Composable (Modifier) -> Unit>(
-        { mod ->
-            SparkTile(
-                mod, "Rest",
-                value = pctValue(m.performance.latest),
-                caption = vsTypical(m.performance.latest, m.performance.typical, "%"),
-                accent = m.performance.latest?.let { Palette.recoveryColor(it) } ?: Palette.textPrimary,
-                spark = m.performance.series, sparkColor = Palette.restColor,
-                onClick = { onMetricClick("performance") },
-            )
-        },
-        { mod ->
-            SparkTile(
-                mod, "Efficiency",
-                value = pctValue(m.efficiency.latest),
-                caption = vsTypical(m.efficiency.latest, m.efficiency.typical, "%"),
-                accent = Palette.statusPositive,
-                spark = m.efficiency.series, sparkColor = Palette.statusPositive,
-                onClick = { onMetricClick("efficiency") },
-            )
-        },
-        { mod ->
-            SparkTile(
-                mod, "Consistency",
-                value = pctValue(m.consistency.latest),
-                caption = vsTypical(m.consistency.latest, m.consistency.typical, "%"),
-                accent = m.consistency.latest?.let { Palette.recoveryColor(it) } ?: Palette.textPrimary,
-                spark = m.consistency.series, sparkColor = Palette.metricCyan,
-                onClick = { onMetricClick("consistency") },
-            )
-        },
-        { mod ->
-            SparkTile(
-                mod, "Hours vs Needed",
-                value = pctValue(m.hoursVsNeeded.latest),
-                caption = vsTypical(m.hoursVsNeeded.latest, m.hoursVsNeeded.typical, "%"),
-                accent = m.hoursVsNeeded.latest?.let { Palette.recoveryColor(minOf(100.0, it)) } ?: Palette.textPrimary,
-                spark = m.hoursVsNeeded.series, sparkColor = Palette.restColor,
-                onClick = { onMetricClick("hours_vs_needed") },
-            )
-        },
-        { mod ->
-            SparkTile(
-                mod, "Restorative",
-                value = pctValue(m.restorative.latest),
-                caption = vsTypical(m.restorative.latest, m.restorative.typical, "%"),
-                accent = Palette.sleepREM,
-                spark = m.restorative.series, sparkColor = Palette.sleepREM,
-                onClick = { onMetricClick("restorative") },
-            )
-        },
-        { mod ->
-            SparkTile(
-                mod, "Respiratory",
-                value = m.respiratory.latest?.let { String.format(Locale.US, "%.1f", it) } ?: "—",
-                caption = vsTypical(m.respiratory.latest, m.respiratory.typical, " rpm", decimals = 1),
-                accent = Palette.metricPurple,
-                spark = m.respiratory.series, sparkColor = Palette.metricPurple,
-                onClick = { onMetricClick("respiratory") },
-            )
-        },
-    )
-
-    Column(verticalArrangement = Arrangement.spacedBy(Metrics.gap)) {
-        SectionHeader("Night detail", overline = "Metrics", trailing = "vs typical")
-
-        // Sleep Debt is the actionable summary for the section, so it leads at the full
-        // two-column width. The remaining six peer metrics keep the established 2 × 3 grid.
-        SparkTile(
-            Modifier.fillMaxWidth(), "Sleep Debt",
-            value = m.sleepDebt.latest?.let { durationText(it) } ?: "—",
-            caption = debtCaption(m.sleepDebt.latest),
-            accent = debtColor(m.sleepDebt.latest),
-            spark = m.sleepDebt.series, sparkColor = Palette.metricRose,
-            onClick = { onMetricClick("sleep_debt") },
-        )
-
-        // Two-up rows; IntrinsicSize.Max + fillMaxHeight keep row neighbors equal height even when
-        // large font scales grow one tile past the tileHeight floor. No empty cells.
-        tiles.chunked(2).forEach { rowTiles ->
-            Row(
-                modifier = Modifier.height(IntrinsicSize.Max),
-                horizontalArrangement = Arrangement.spacedBy(Metrics.gap),
-            ) {
-                rowTiles.forEach { it(Modifier.weight(1f).fillMaxHeight()) }
-            }
-        }
-    }
-}
-
-// MARK: - 2b. Sleep-debt ledger (rolling 14-night running balance)
-
-/**
- * A running balance of (slept − personal need) across the recent fortnight, surfaced as one
- * card: the net debt/surplus headline, a plain-English read, and a diverging bar of each
- * night's delta (surplus above the centre line, deficit below). Honest: a simple accumulator
- * — a surplus night offsets a deficit one — capped at 14 nights, no-data nights skipped.
- * Mirrors the macOS SleepView sleepDebtLedger card section-for-section. (#242)
- */
-@Composable
-internal fun SleepDebtLedgerCard(ledger: SleepDebtLedger) {
-    Column(verticalArrangement = Arrangement.spacedBy(Metrics.gap)) {
-        SectionHeader("Sleep-debt ledger", overline = "Last 14 nights", trailing = "running balance")
-        NoopCard(padding = Metrics.cardPadding, tint = Palette.restColor) {
-            if (ledger.nightCount == 0) {
-                Text(
-                    uiString(R.string.l10n_sleep_screen_no_nights_with_sleep_data_yet_fa71b6b3),
-                    style = NoopType.subhead,
-                    color = Palette.textTertiary,
-                )
-            } else {
-                Column(verticalArrangement = Arrangement.spacedBy(Metrics.space14)) {
-                    // Headline: net balance + the short tag (sleep debt / surplus / balanced).
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text(
-                            debtHeadline(ledger),
-                            style = NoopType.tileValueLarge,
-                            color = debtBalanceColor(ledger),
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis,
-                            modifier = Modifier.weight(1f),
-                        )
-                        Text(
-                            debtTag(ledger),
-                            style = NoopType.captionNumber,
-                            color = debtBalanceColor(ledger),
-                        )
-                    }
-                    // Plain-English read.
-                    Text(
-                        debtRead(ledger),
-                        style = NoopType.subhead,
-                        color = Palette.textSecondary,
-                    )
-                    // Per-night diverging delta bars (surplus up, deficit down).
-                    DebtDeltaBars(ledger)
-                    Hairline()
-                    ChartFooter(
-                        listOf(
-                            "Balance" to debtSigned(ledger.balanceMin),
-                            "Per-night need" to durationText(ledger.needMin),
-                            "Nights" to "${ledger.nightCount}",
-                        ),
-                    )
-                }
-            }
-        }
-    }
-}
-
-/**
- * The diverging per-night delta strip: each night a bar from the centre line — up (accent)
- * for a surplus, down (rose) for a deficit — scaled to the largest |delta|.
- */
-@Composable
-private fun DebtDeltaBars(ledger: SleepDebtLedger) {
-    val deltas = ledger.nights.map { it.deltaMin }
-    val scale = max(deltas.maxOfOrNull { abs(it) } ?: 1.0, 1.0)
-    val accentColor = Palette.accent
-    val deficitColor = Palette.metricRose
-    val centreColor = Palette.hairline
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(56.dp)
-            .semantics {
-                contentDescription =
-                    uiString(R.string.l10n_sleep_screen_per_night_sleep_balance_ledger_nightcount_f339d0ab, ledger.nightCount, debtSigned(ledger.balanceMin))
-            }
-            .drawBehind {
-                val n = max(deltas.size, 1)
-                val slot = size.width / n
-                val barW = max(2f, slot * 0.6f)
-                val midY = size.height / 2f
-                // Centre (zero) line.
-                drawLine(
-                    color = centreColor,
-                    start = Offset(0f, midY),
-                    end = Offset(size.width, midY),
-                    strokeWidth = 1f,
-                )
-                deltas.forEachIndexed { i, d ->
-                    val frac = (abs(d) / scale).toFloat().coerceIn(0f, 1f)
-                    val h = max(2f, frac * (midY - 2f))
-                    val cx = slot * i + slot / 2f
-                    // Surplus grows upward from the centre, deficit downward.
-                    val top = if (d >= 0.0) midY - h else midY
-                    drawRoundRect(
-                        color = if (d >= 0.0) accentColor else deficitColor,
-                        topLeft = Offset(cx - barW / 2f, top),
-                        size = Size(barW, h),
-                        cornerRadius = CornerRadius(2f, 2f),
-                    )
-                }
-            },
-    )
-}
-
-// MARK: - 3. Stages vs typical
-
-@Composable
-private fun StagesVsTypical(m: SleepModel) {
-    val s = m.stages
-    Column(verticalArrangement = Arrangement.spacedBy(Metrics.gap)) {
-        SectionHeader("Stages vs typical", overline = "Selected night", trailing = "marker = your mean")
-        NoopCard(tint = Palette.restColor) {
-            Column(verticalArrangement = Arrangement.spacedBy(Metrics.space14)) {
-                StageRow("Deep", last = s.deep, typical = m.typicalDeepMin, color = Palette.sleepDeep)
-                Hairline()
-                StageRow("REM", last = s.rem, typical = m.typicalRemMin, color = Palette.sleepREM)
-                Hairline()
-                StageRow("Light", last = s.light, typical = m.typicalLightMin, color = Palette.sleepLight)
-            }
-        }
-    }
-}
-
-@Composable
-private fun Hairline() {
-    Box(modifier = Modifier.fillMaxWidth().height(Metrics.divider).background(Palette.hairline))
-}
-
-/** One stage bar: last-night minutes filled, with a vertical marker at the typical mean. */
-@Composable
-private fun StageRow(label: String, last: Double, typical: Double?, color: Color) {
-    val scaleMax = max(last, typical ?: 0.0) * 1.18
-    val scale = if (scaleMax > 0.0) scaleMax else 1.0
-    val deltaText: String = run {
-        if (typical == null || typical <= 0.0) {
-            ""
-        } else {
-            val diff = last - typical
-            val sign = if (diff >= 0) "+" else "−"
-            "$sign${durationText(abs(diff))} vs typ"
-        }
-    }
-    Column(verticalArrangement = Arrangement.spacedBy(Metrics.space6)) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Overline(label, modifier = Modifier.weight(1f))
-            Text(durationText(last), style = NoopType.captionNumber, color = Palette.textPrimary)
-            if (deltaText.isNotEmpty()) {
-                Text(
-                    deltaText,
-                    style = NoopType.footnote,
-                    color = if (last >= (typical ?: last)) Palette.statusPositive else Palette.statusWarning,
-                    modifier = Modifier.padding(start = Metrics.space8),
-                )
-            }
-        }
-        // Track + last-night fill + typical marker.
-        val fillFrac = (last / scale).coerceIn(0.0, 1.0).toFloat()
-        val markerFrac = typical?.takeIf { it > 0.0 }?.let { (it / scale).coerceIn(0.0, 1.0).toFloat() }
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(Metrics.progressHeight)
-                .clip(RoundedCornerShape(Metrics.cornerPill))
-                .background(Palette.surfaceInset)
-                .semantics { contentDescription = uiString(R.string.l10n_sleep_screen_label_minutes_vs_your_typical_bar_b8f6a482, label) }
-                .drawBehind {
-                    // last-night fill
-                    if (fillFrac > 0f) {
-                        drawRoundRectFill(color, fillFrac)
-                    }
-                    // typical marker
-                    if (markerFrac != null) {
-                        val x = (size.width * markerFrac).coerceIn(1f, size.width - 1f)
-                        drawLine(
-                            color = Palette.textPrimary,
-                            start = Offset(x, 0f),
-                            end = Offset(x, size.height),
-                            strokeWidth = 2f,
-                            cap = StrokeCap.Round,
-                        )
-                    }
-                },
-        )
-    }
-}
-
-private fun DrawScope.drawRoundRectFill(color: Color, frac: Float) {
-    val w = (size.width * frac).coerceAtLeast(size.height)
-    val r = size.height / 2f
-    drawRoundRect(
-        color = color,
-        size = Size(w, size.height),
-        cornerRadius = CornerRadius(r, r),
-    )
-}
-
-// MARK: - 4. 14-day asleep-hours trend
-
-@Composable
-private fun DurationTrend(m: SleepModel) {
-    val pts = m.trendHours
-    val avg = pts.sleepAverageOrNull()
-    Column(verticalArrangement = Arrangement.spacedBy(Metrics.gap)) {
-        SectionHeader("Trend", overline = "Sleep", trailing = "Last 14 days")
-        ChartCard(
-            title = uiString(R.string.l10n_sleep_screen_hours_asleep_06f68993),
-            subtitle = "Per night, trailing 14 days",
-            trailing = avg?.let { String.format(Locale.US, "%.1f h avg", it) },
-            tint = Palette.restColor,
-            footer = {
-                ChartFooter(
-                    listOf(
-                        "Avg" to (avg?.let { String.format(Locale.US, "%.1f h", it) } ?: "—"),
-                        "Min" to (pts.minOrNull()?.let { String.format(Locale.US, "%.1f h", it) } ?: "—"),
-                        "Max" to (pts.maxOrNull()?.let { String.format(Locale.US, "%.1f h", it) } ?: "—"),
-                        "Nights" to "${pts.size}",
-                    ),
-                )
-            },
-        ) {
-            if (pts.size >= 2) {
-                Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-                    // #85: sleep duration reads as a per-night histogram (zero-based bars), matching the
-                    // iOS Sleep tab's TrendChart(showsBars:) — a BarMark is proportional to hours slept,
-                    // clearer than a line for a nightly total. BarChart floors at 0 like the iOS bar domain.
-                    BarChart(
-                        values = pts,
-                        modifier = Modifier.fillMaxWidth().height(Metrics.compactChartHeight)
-                            .semantics { contentDescription = uiString(R.string.l10n_sleep_screen_sleep_hours_trend_chart_a6fbc46d) },
-                        color = Palette.restColor,
-                        selectionEnabled = true,
-                        // #691: on tap, show the DATE alongside the value (the shared chart's tooltip),
-                        // matching the other trend graphs. trendDates is index-aligned with the values.
-                        selectionLabels = m.trendDates.map(::shortDayLabel),
-                    )
-                    DateAxisRow(m.trendDates)
-                }
-            } else {
-                TrendPlaceholder()
-            }
-        }
-
-        ChartCard(
-            title = uiString(R.string.l10n_sleep_screen_sleep_debt_3aec7d9c),
-            subtitle = "Sleep debt per day",
-            // #691: sleep debt is usually well under an hour, so decimal hours ("0.6h") reads badly —
-            // show hours+minutes. trendDebtHours is in hours; durationText takes minutes.
-            trailing = m.trendDebtHours.lastOrNull()?.let { durationText(it * 60.0) },
-            tint = Palette.restColor,
-            footer = {
-                ChartFooter(
-                    listOf(
-                        "Avg" to (m.trendDebtHours.sleepAverageOrNull()?.let { durationText(it * 60.0) } ?: "â€”"),
-                        "Max" to (m.trendDebtHours.maxOrNull()?.let { durationText(it * 60.0) } ?: "â€”"),
-                        "Days" to "${m.trendDebtHours.size}",
-                    ),
-                )
-            },
-        ) {
-            if (m.trendDebtHours.isNotEmpty()) {
-                Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-                    BarChart(
-                        values = m.trendDebtHours,
-                        modifier = Modifier.fillMaxWidth().height(Metrics.compactChartHeight)
-                            .semantics { contentDescription = uiString(R.string.l10n_sleep_screen_sleep_debt_trend_chart_9e178776) },
-                        color = Palette.metricRose,
-                        selectionEnabled = true,
-                        selectionLabels = m.trendDates.map(::shortDayLabel),   // #691: hover shows date + value
-                    )
-                    DateAxisRow(m.trendDates)
-                }
-            } else {
-                TrendPlaceholder()
-            }
-        }
-    }
-}
-
-@Composable
-private fun TrendPlaceholder() {
-    Box(
-        modifier = Modifier.fillMaxWidth(),
-        contentAlignment = Alignment.Center,
-    ) {
-        InsetChartPlaceholder(message = "Not enough nights yet.")
-    }
-}
-
-@Composable
-private fun TrendLegend(items: List<Pair<String, Color>>) {
-    Row(horizontalArrangement = Arrangement.spacedBy(Metrics.space14)) {
-        items.forEach { (label, color) ->
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(Metrics.space6),
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Box(
-                    modifier = Modifier
-                        .width(Metrics.legendLineWidth)
-                        .height(Metrics.legendLineHeight)
-                        .clip(RoundedCornerShape(Metrics.cornerPill))
-                        .background(color),
-                )
-                Text(label, style = NoopType.footnote, color = Palette.textTertiary)
-            }
-        }
-    }
-}
-
-@Composable
-private fun DateAxisRow(days: List<String>) {
-    if (days.isEmpty()) return
-    val labels = listOf(
-        days.firstOrNull(),
-        days.getOrNull(days.lastIndex / 2),
-        days.lastOrNull(),
-    ).map { it?.let(::shortDayLabel).orEmpty() }
-    Row(modifier = Modifier.fillMaxWidth()) {
-        labels.forEach { label ->
-            Text(
-                text = label,
-                style = NoopType.footnote,
-                color = Palette.textTertiary,
-                modifier = Modifier.weight(1f),
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-    }
-}
-
-// MARK: - ChartCard / ChartFooter (local — mirror the macOS ChartCard the screen used)
-
-/**
- * The chart container the macOS screen leaned on: a NoopCard with a header (overline-
- * style title + subtitle + trailing read-out), the chart body, then a footer row of
- * label/value pairs. Kept local so the shared component set stays minimal.
- */
-@Composable
-private fun ChartCard(
-    title: String,
-    subtitle: String,
-    trailing: String?,
-    footer: @Composable () -> Unit,
-    tint: Color? = null,
-    chart: @Composable () -> Unit,
-) {
-    NoopCard(padding = Metrics.cardPadding, tint = tint) {
-        Column(verticalArrangement = Arrangement.spacedBy(Metrics.space14)) {
-            Row(verticalAlignment = Alignment.Top) {
-                Column(modifier = Modifier.weight(1f)) {
-                    Text(title, style = NoopType.headline, color = Palette.textPrimary)
-                    Text(subtitle, style = NoopType.footnote, color = Palette.textSecondary)
-                }
-                if (trailing != null) {
-                    Text(trailing, style = NoopType.chartValue, color = Palette.textPrimary)
-                }
-            }
-            chart()
-            footer()
-        }
-    }
-}
-
-/** A footer strip of label/value pairs, evenly distributed. */
-@Composable
-private fun ChartFooter(items: List<Pair<String, String>>) {
-    Row(modifier = Modifier.fillMaxWidth()) {
-        items.forEach { (label, value) ->
-            Column(modifier = Modifier.weight(1f)) {
-                Overline(label, color = Palette.textTertiary)
-                // Stage-breakdown values like "1h 23m (24%)" wrapped to a second line in a narrow column,
-                // pushing the row taller and clipping against the card edge (#406). Hold them to one line.
-                Text(
-                    value,
-                    style = NoopType.captionNumber,
-                    color = Palette.textPrimary,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    softWrap = false,
-                )
-            }
-        }
-    }
-}
-
-// MARK: - SparkTile (min-height metric tile, stacked: value + caption over a full-width 30-day sparkline)
-
-@Composable
-private fun SparkTile(
-    modifier: Modifier,
-    label: String,
-    value: String,
-    caption: String?,
-    accent: Color,
-    spark: List<Double>,
-    sparkColor: Color,
-    onClick: (() -> Unit)? = null,
-) {
-    // liquidPress on the tappable tile: it settles inward on press (the pilot's card feel). The SAME
-    // interactionSource drives the clickable + the press; indication = null so only the liquid settle shows.
-    val interaction = remember { MutableInteractionSource() }
-    // heightIn (not height): tileHeight is a floor, matching the Swift StatTile. At normal font scale the
-    // tile keeps its 108dp footprint; at large font scales it grows instead of clipping the caption. (#squish)
-    val clickMod = if (onClick != null) {
-        modifier
-            .heightIn(min = Metrics.tileHeight)
-            .liquidPress(interaction)
-            .clickable(interactionSource = interaction, indication = null, onClick = onClick)
-    } else {
-        modifier.heightIn(min = Metrics.tileHeight)
-    }
-    NoopCard(modifier = clickMod, padding = Metrics.space14) {
-        // fillMaxHeight so the weight-spacer can pin the sparkline to the card bottom once the
-        // MetricGrid row bounds the height (Row height(IntrinsicSize.Max) + tile fillMaxHeight()).
-        Column(modifier = Modifier.fillMaxWidth().fillMaxHeight()) {
-            Overline(label)
-            Text(
-                value,
-                style = NoopType.tileValue,
-                color = accent,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-            if (caption != null) {
-                Text(
-                    caption,
-                    style = NoopType.footnote,
-                    color = Palette.textTertiary,
-                    // Full card width now, so the "-3% vs typical" caption fits; ellipsis stays as a
-                    // safety net for extreme localized strings.
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier.padding(top = Metrics.space2),
-                )
-            }
-            Spacer(Modifier.weight(1f))
-            val tail = spark.takeLast(30)
-            if (tail.size >= 2) {
-                // Full-width bottom spark. Outer height(sparkHeight) deliberately overrides Sparkline's
-                // internal 28dp default down to the 22dp tile spark (same override SparkTailBox does).
-                Sparkline(
-                    values = tail,
-                    color = sparkColor,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(top = Metrics.space8)
-                        .height(Metrics.sparkHeight),
-                )
-            }
-        }
-    }
-}
+// MARK: - Sleep metric cards and charts UI lives in SleepMetricCardsUi.kt
 
 // MARK: - Empty state
 
@@ -2786,7 +1738,7 @@ internal fun HoursVsNeededCard(m: SleepModel) {
                 LegendDot("Debt", Palette.statusCritical)
             }
 
-            Hairline()
+            SleepMetricsHairline()
             Row(modifier = Modifier.fillMaxWidth()) {
                 listOf(
                     "Slept" to String.format(Locale.US, "%.1f h", sleptH),
@@ -2972,7 +1924,7 @@ internal fun SleepConsistencyCard(sleeps: List<SleepSession>, habitualMidsleepSe
                 LegendDot("Wake  $typicalWakeLabel", Palette.restColor)
             }
 
-            Hairline()
+            SleepMetricsHairline()
             Row(modifier = Modifier.fillMaxWidth()) {
                 listOf(
                     "Score" to "${consistencyPct.roundToInt()}%",
@@ -2989,112 +1941,4 @@ internal fun SleepConsistencyCard(sleeps: List<SleepSession>, habitualMidsleepSe
     }
 }
 
-// MARK: - Sleep metric detail sheet
-
-@Composable
-private fun SleepMetricDetailSheetContent(vm: AppViewModel, key: String) {
-    val days by vm.recentDays.collectAsStateWithLifecycle()
-    var range by remember { mutableStateOf(SleepMetricRange.MONTH) }
-    val spec = remember(key) { sleepMetricSpec(key) }
-    val allPoints = remember(days, key) { buildSleepMetricPoints(days, key) }
-    val filteredPoints = remember(allPoints, range) { filterSleepMetricPoints(allPoints, range) }
-
-    Column(
-        modifier = Modifier.fillMaxWidth().padding(horizontal = Metrics.space24, vertical = Metrics.space8),
-        verticalArrangement = Arrangement.spacedBy(Metrics.space16),
-    ) {
-        if (allPoints.size < 2) {
-            Text(uiString(R.string.l10n_sleep_screen_not_enough_history_yet_0e2f93b6), style = NoopType.headline, color = Palette.textPrimary)
-            Text(
-                uiString(R.string.l10n_sleep_screen_this_metric_needs_at_least_two_2de1d37a),
-                style = NoopType.subhead, color = Palette.textSecondary,
-            )
-            Spacer(Modifier.height(Metrics.space16))
-        } else if (filteredPoints.size < 2) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Column(modifier = Modifier.weight(1f)) {
-                    Overline("Sleep")
-                    Text(spec.title, style = NoopType.title2, color = Palette.textPrimary)
-                }
-            }
-            SegmentedPillControl(
-                items = SleepMetricRange.entries,
-                selection = range,
-                label = { it.label },
-                onSelect = { range = it },
-            )
-            Text(uiString(R.string.l10n_sleep_screen_not_enough_history_in_this_range_7e2fd640), style = NoopType.subhead, color = Palette.textSecondary)
-            Spacer(Modifier.height(Metrics.space16))
-        } else {
-            val values = filteredPoints.map { it.second }
-            val dates = filteredPoints.map { it.first }
-            val latest = filteredPoints.last()
-            val minV = values.minOrNull() ?: 0.0
-            val maxV = values.maxOrNull() ?: 0.0
-            val avgV = values.average()
-
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Column(modifier = Modifier.weight(1f)) {
-                    Overline("Sleep · ${filteredPoints.size} nights")
-                    Text(spec.title, style = NoopType.title2, color = Palette.textPrimary)
-                    Text(uiString(R.string.l10n_sleep_screen_as_of_latest_first_726f20bb, latest.first), style = NoopType.footnote, color = Palette.textTertiary)
-                }
-                Text(
-                    uiString(R.string.l10n_sleep_screen_spec_format_latest_second_spec_unit_18433019, spec.format(latest.second), spec.unit).trim(),
-                    style = NoopType.chartValue,
-                    color = spec.color,
-                )
-            }
-            SegmentedPillControl(
-                items = SleepMetricRange.entries,
-                selection = range,
-                label = { it.label },
-                onSelect = { range = it },
-            )
-            Row(
-                modifier = Modifier.height(IntrinsicSize.Min),
-                horizontalArrangement = Arrangement.spacedBy(Metrics.space4),
-            ) {
-                Column(
-                    modifier = Modifier.height(Metrics.chartHeight),
-                    verticalArrangement = Arrangement.SpaceBetween,
-                ) {
-                    Text(uiString(R.string.l10n_sleep_screen_spec_format_maxv_spec_unit_65091104, spec.format(maxV), spec.unit).trim(), style = NoopType.footnote, color = Palette.textTertiary, maxLines = 1)
-                    Text(uiString(R.string.l10n_sleep_screen_spec_format_avgv_spec_unit_46bf7fdc, spec.format(avgV), spec.unit).trim(), style = NoopType.footnote, color = Palette.textTertiary, maxLines = 1)
-                    Text(uiString(R.string.l10n_sleep_screen_spec_format_minv_spec_unit_e69978f4, spec.format(minV), spec.unit).trim(), style = NoopType.footnote, color = Palette.textTertiary, maxLines = 1)
-                }
-                LineChart(
-                    values = values,
-                    modifier = Modifier.weight(1f).height(Metrics.chartHeight)
-                        .semantics { contentDescription = uiString(R.string.l10n_sleep_screen_spec_title_trend_chart_3085ac6e, spec.title) },
-                    color = spec.color,
-                    fill = true,
-                    selectionEnabled = true,
-                    selectionLabels = filteredPoints.map { shortDayLabel(it.first) },   // #691: hover shows date + value
-                )
-            }
-            Row(modifier = Modifier.fillMaxWidth()) {
-                listOf(dates.first(), dates.getOrNull(dates.lastIndex / 2), dates.last()).forEach { d ->
-                    Text(
-                        d?.let { runCatching { LocalDate.parse(it).format(DateTimeFormatter.ofPattern("d MMM", Locale.US)) }.getOrDefault(it) }.orEmpty(),
-                        style = NoopType.footnote, color = Palette.textTertiary,
-                        modifier = Modifier.weight(1f), maxLines = 1, overflow = TextOverflow.Ellipsis,
-                    )
-                }
-            }
-            Hairline()
-            Row(modifier = Modifier.fillMaxWidth()) {
-                listOf("Min" to minV, "Avg" to avgV, "Max" to maxV).forEach { (lbl, v) ->
-                    Column(modifier = Modifier.weight(1f)) {
-                        Overline(lbl, color = Palette.textTertiary)
-                        Text(
-                            uiString(R.string.l10n_sleep_screen_spec_format_v_spec_unit_7a7f630c, spec.format(v), spec.unit).trim(),
-                            style = NoopType.captionNumber, color = Palette.textPrimary,
-                        )
-                    }
-                }
-            }
-            Spacer(Modifier.height(Metrics.space8))
-        }
-    }
-}
+// MARK: - Sleep metric detail sheet UI lives in SleepMetricDetailSheet.kt
