+++ android/app/src/main/res/values-pt-rPT/strings.xml
@@ -1715,6 +1715,8 @@
     </plurals>
     <string name="timeline_spo2_not_on_whoop5">Esta bracelete não envia SpO₂ por Bluetooth. Importe uma exportação WHOOP ou Health Connect para a ver.</string>
     <string name="timeline_resp_not_on_whoop5">Esta bracelete não envia qualquer fluxo respiratório bruto. A tua frequência respiratória estimada aparece no ecrã Saúde.</string>
+    <string name="sleep_delete_user_edited_body">Remove este sono e recalcula o dia sem ele. Podes anular durante alguns segundos.</string>
+    <string name="sleep_delete_recorded_body">Remove este sono registado e recalcula o dia sem ele. O NOOP não vai voltar a detetar sono nesta janela. Podes anular durante alguns segundos.</string>
     <string name="l10n_data_sources_screen_last_shared_6bf8389c">"Última partilha "</string>
     <string name="l10n_data_sources_screen_sharing_paused_health_connect_permission_was_e8950315">Partilha pausada: a permissão do Health Connect foi revogada. Toca para conceder novamente.</string>
     <string name="l10n_data_sources_screen_last_share_didn_t_finish_noop_0d6c27f0">A última partilha não foi concluída – o NOOP tentará novamente automaticamente.</string>
