+++ android/app/src/main/res/values-zh/strings.xml
@@ -1658,6 +1658,8 @@
     <string name="explore_description_rest">你的睡眠质量与修复力：综合睡眠时长、睡眠效率、深睡+REM占比以及入睡时机等。</string>
     <string name="timeline_spo2_not_on_whoop5">此设备不通过蓝牙发送 SpO₂。导入 WHOOP 数据或 Health Connect 即可查看。</string>
     <string name="timeline_resp_not_on_whoop5">此设备不发送原始呼吸数据流。你的估算呼吸频率显示在“健康”屏幕上。</string>
+    <string name="sleep_delete_user_edited_body">移除此次睡眠并重新计算当天数据。之后几秒内可以撤销。</string>
+    <string name="sleep_delete_recorded_body">移除此次已记录的睡眠并重新计算当天数据。NOOP 不会在此时间段内重新检测睡眠。之后几秒内可以撤销。</string>
     <string name="l10n_data_sources_screen_last_shared_6bf8389c">"上次分享 "</string>
     <string name="l10n_data_sources_screen_sharing_paused_health_connect_permission_was_e8950315">分享已暂停——Health Connect 权限已被撤销。点按以重新授予。</string>
     <string name="l10n_data_sources_screen_last_share_didn_t_finish_noop_0d6c27f0">上次分享未完成——NOOP 将自动重试。</string>
